requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Authlib==1.2.0
Werkzeug==3.0.0
PyJWT==2.6.0